Reland "[wasm-c-api][test] Run Wasm C API tests on bots"
by including them in the "bot_default" and "default" test sets.
The build targets are already up to date, only the test runner
needs to be updated.

This is a reland of a5d279da00738fe648e1341d81c116f771ec0dd7.
This CL itself is unchanged since the previous attempt; the
issue that caused the revert was fixed by
https://chromium-review.googlesource.com/c/v8/v8/+/1781695.

Originally reviewed on:
https://chromium-review.googlesource.com/c/v8/v8/+/1768370

Change-Id: Ifd40a398b8d55ff948f5acfb25cb676937cb7794
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/1789158
Reviewed-by: Tamer Tas <tmrts@chromium.org>
Commit-Queue: Jakob Kummerow <jkummerow@chromium.org>
Cr-Commit-Position: refs/heads/master@{#63619}